+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - ICollisionable interface
 - BoxInteractor component
 - SphereInteractor component
+- CollisionableUnityEvent component
 
 ## [1.0.0] - 2025-10-30
 ### Added